:sparkles: [feat] 단건 포스트 조회 응답 dto의 location 양식 수정 - PostDetailResponse.java - PostMapper.java - LocationDto.java - PostDetailResult.java - PostQueryAdapter.java

+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/post/mapper/PostMapper.java
@@ -40,11 +40,12 @@ public class PostMapper {
     }
 
     public PostDetailResponse toDetailResponse(PostDetailResult result) {
+
         return PostDetailResponse.builder()
                 .title(result.getTitle())
                 .content(result.getContent())
-                .rentalFee(result.getRentalFee())
-                .deposit(result.getDeposit())
+                .rentalFee((long) result.getRentalFee())
+                .deposit((long) result.getDeposit())
                 .address(result.getAddress())
                 .location(result.getLocation())
                 .productImage(result.getProductImages())
